Convert redux/reducer/redirects to typescript
